+++ layout/CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+* Use font line_gap as top-padding for layouts to address issues with glyphs taller than the font ascent.
+
 # 0.2.3
 * Default layouts: Keep word trailing space width if ending in a hard break or end of all glyphs _e.g. `"Foo  \n"`_ _(This particularly changes the layout of right & centre aligned text ending in spaces)_.